+++ routes/web.php
@@ -24,34 +24,37 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [ListingController::class, 'index']);
 
 // Show Create Form
-Route::get('listings/create', [ListingController::class, 'create']);
+Route::get('listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing Date
-Route::post('listings', [ListingController::class, 'store']);
+Route::post('listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show Edit Form
-Route::get('listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Listing
-Route::put('listings/{listing}', [ListingController::class, 'update']);
+Route::put('listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete Listing
-Route::delete('listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Single Listing
 Route::get('listings/{listing}', [ListingController::class, 'show'])->where('id', '[0-9]+');
 
 // Show Register/Create Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Create New User
 Route::post('/users', [UserController::class, 'store']);
 
 // Log User Out
-Route::post('logout', [UserController::class, 'logout']);
+Route::post('logout', [UserController::class, 'logout'])->middleware('auth');
 
 // Show Login Form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Log In User
 Route::post('users/authenticate', [UserController::class, 'authenticate']);
